- front - data i czas ostatniej bitwy - winrate - pobieranie statystyk gracza TBD: - inne statystyki ??? - WR z przedzialu czasu - liczenie WR od danego momentu - front-end

+++ src/main/java/pl/lukabrasi/wotblitzstats/dtos/AllDto.java
@@ -7,7 +7,7 @@ public class AllDto {
 
     private double wins;
     private double losses;
-    private double battles;
+    private int battles;
     private double winrate;
 
     public double getWinrate() {
